% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliationgraph.R
\name{affiliation_graph}
\alias{affiliation_graph}
\title{Visualize an affiliation network}
\usage{
affiliation_graph(edgelist)
}
\arguments{
\item{edgelist}{a dataframe that contains a list of people and their associated activities. The first column
should list the name of the source node; the second column should list associated activities.}
}
\description{
Function for plotting affiliation networks.
}
\examples{
# minimal example
agents <-  c("Bill", "Bill", "Bill", "Bob", "Bob", "Bob", "John", "John")
activities <- c("Act1","Act2", "Act3","Act1", "Act2","Act4","Act3","Act4")
affiliations <- data.frame(agents, activities)
affiliation_graph(edgelist = affiliations)

# FM 3-24 example affiliation matrix
data("fm3_24_affiliation")
affiliations <- as.data.frame(fm3_24_affiliation)
affiliation_graph(edgelist = affiliations)

}
