% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralitytable.R
\name{centralitytable}
\alias{centralitytable}
\title{Create a table of centrality values}
\usage{
centralitytable(edgelist)
}
\arguments{
\item{edgelist}{a dataframe that contains a list of people and their associates. The first column represents
source nodes and the second column represents target nodes.}
}
\description{
Function for generating a table of the nodes ranked by four common centrality measures (degree,
eigenvector, betweenness, and closeness centrality)
}
\details{
Degree centrality identifies well-connected nodes.
Eigenvector centrality identifies nodes that are connected to well-connected nodes.
Betweenness centrality identifies nodes that serve as bridges.
Closeness centrality identifies nodes that are close to all other nodes in a network.
}
\examples{
# minimal example
source <- c("a", "a", "b", "c", "d", "d", "d")
target <- c("b", "c", "c", "d", "e", "f", "g")
associations <- data.frame(source, target)
centralitytable(edgelist = associations)

# FM 3-24 example association matrix
data("fm3_24_affiliations")
affiliations <- as.data.frame(fm3_24_affiliations)
associations <- transform2social(edgelist = affiliations)
centralitytable(edgelist = associations)

}
