% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communitytable.R
\name{communitytable}
\alias{communitytable}
\title{Create a table of structurally based communities}
\usage{
communitytable(edgelist)
}
\arguments{
\item{edgelist}{a dataframe that contains a list of people and their associates. The first column represents
source nodes and the second column represents target nodes.}
}
\description{
Function for generating a table of the communities generated by the Girvan-Newman edge betweenness community
detection algorithm.
}
\examples{
# minimal example
source <- c("a", "a", "b", "c", "d", "d", "d")
target <- c("b", "c", "c", "d", "e", "f", "g")
associations <- data.frame(source, target)
communitytable(edgelist = associations)

# FM 3-24 example association matrix
data("fm3_24_affiliations")
affiliations <- as.data.frame(fm3_24_affiliations)
associations <- transform2social(edgelist = affiliations)
communitytable(edgelist = associations)

}
