% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphcommunities.R
\name{graph_communities}
\alias{graph_communities}
\title{Community detection using edge betweenness}
\usage{
graph_communities(edgelist, sizeby = "none", weighted = FALSE)
}
\arguments{
\item{sizeby}{custom option to size nodes based on either "degree", "eigenvector", "betweenness", or "closeness" centrality.
Default to "none".}

\item{weighted}{custom option to size edges based on weighted values. Default to "FALSE".}

\item{edglist}{a dataframe that contains a list of people and their associates. The first column represents
source nodes and the second column represents target nodes.}
}
\description{
Function that identifies structurally-based groups using the Girvan–Newman, edge betweenness community detection algorithm
and then produces a plot where nodes are colored based on group membership.
}
\examples{
# minimal example
source <- c("a", "a", "b", "c", "d", "d", "d")
target <- c("b", "c", "c", "d", "e", "f", "g")
associations <- data.frame(source, target)
graph_communities(edgelist = associations, sizeby = "betweenness")

# FM 3-24 example affiliation matrix
data("fm3_24_affiliation")
affiliations <- as.data.frame(fm3_24_affiliation)
associations <- transform2social(edgelist = affiliations)
graph_communities(edgelist = associations, sizeby = "betweenness", weighted = TRUE)

}
